refactoring Traffic Handler for passing messages

+++ src/main/java/eu/uberdust/testbedlistener/util/HereIamMessage.java
@@ -15,7 +15,7 @@ public class HereIamMessage {
         return mess;
     }
 
-    public HereIamMessage(int[] payload) {
+    public HereIamMessage(byte[] payload) {
 
         StringBuilder sb = new StringBuilder();
 //        System.out.println(payload.length);
